sc-dev: WIP Har reader
