+++ hedra/plugins/types/extension/types.py
@@ -1,0 +1,5 @@
+from enum import Enum
+
+
+class ExtensionType(Enum):
+    GENERATOR='GENERATOR'
